+++ src/webview/index.ts
@@ -1,4 +1,5 @@
 export { IgnorePanel } from "./ignorePanel";
 export { MainPanel } from "./mainPanel";
+export { RequirementsPanel } from "./requirementsPanel";
 export { SettingsPanel } from "./settingsPanel";
 export { StatisticsPanel } from "./statisticsPanel";